Internal refactor: move TPU lowering rules out of jax/_src/lax
