+++ jax/_src/tpu/linalg/stack.py
