+++ tests/stack_test.py
@@ -16,7 +16,7 @@ from absl.testing import absltest
 
 import jax
 import jax.numpy as jnp
-from jax._src.lax.stack import Stack
+from jax._src.tpu.linalg.stack import Stack
 from jax._src import test_util as jtu
